+++ Coimbra/Utilities/GameObjectUtility.cs
@@ -34,37 +34,25 @@ namespace Coimbra
         }
 
         /// <summary>
-        /// Gets the <see cref="GameObjectBehaviour"/> or creates a new default one if missing.
+        /// Creates and initializes a new default <see cref="GameObjectBehaviour"/>.
         /// </summary>
-        public static GameObjectBehaviour GetOrCreateBehaviour(this Component component)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static GameObjectBehaviour CreateBehaviour(this GameObject gameObject)
         {
-            return GetOrCreateBehaviour<GameObjectBehaviour>(component.gameObject);
+            GameObjectBehaviour behaviour = gameObject.AddComponent<GameObjectBehaviour>();
+            behaviour.Initialize();
+
+            return behaviour;
         }
 
         /// <summary>
-        /// Gets the <see cref="GameObjectBehaviour"/> or creates a new default one if missing.
+        /// Creates and initializes a new <see cref="GameObjectBehaviour"/> of the specified type.
         /// </summary>
-        public static T GetOrCreateBehaviour<T>(this Component component)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static T CreateBehaviour<T>(this GameObject gameObject)
             where T : GameObjectBehaviour
         {
-            return GetOrCreateBehaviour<T>(component.gameObject);
-        }
-
-        /// <summary>
-        /// Gets the <see cref="GameObjectBehaviour"/> or creates a new default one if missing.
-        /// </summary>
-        public static GameObjectBehaviour GetOrCreateBehaviour(this GameObject gameObject)
-        {
-            if (Behaviours.TryGetValue(gameObject, out GameObjectBehaviour behaviour))
-            {
-                return behaviour;
-            }
-
-            if (!gameObject.TryGetComponent(out behaviour))
-            {
-                behaviour = gameObject.AddComponent<GameObjectBehaviour>();
-            }
-
+            T behaviour = gameObject.AddComponent<T>();
             behaviour.Initialize();
 
             return behaviour;
@@ -73,29 +61,71 @@ namespace Coimbra
         /// <summary>
         /// Gets the <see cref="GameObjectBehaviour"/> or creates a new default one if missing.
         /// </summary>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static GameObjectBehaviour GetOrCreateBehaviour(this GameObject gameObject)
+        {
+            return TryGetBehaviour(gameObject, out GameObjectBehaviour behaviour) ? behaviour : CreateBehaviour(gameObject);
+        }
+
+        /// <summary>
+        /// Gets the <see cref="GameObjectBehaviour"/> or creates a new one of the specified type if missing.
+        /// </summary>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static T GetOrCreateBehaviour<T>(this GameObject gameObject)
             where T : GameObjectBehaviour
         {
-            if (Behaviours.TryGetValue(gameObject, out GameObjectBehaviour behaviour))
+            return TryGetBehaviour(gameObject, out GameObjectBehaviour behaviour) ? behaviour as T : CreateBehaviour<T>(gameObject);
+        }
+
+        /// <summary>
+        /// Tries to get an existing <see cref="GameObjectBehaviour"/>.
+        /// </summary>
+        /// <returns>True if found a <see cref="GameObjectBehaviour"/>.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static bool TryGetBehaviour(this GameObject gameObject, out GameObjectBehaviour behaviour)
+        {
+            if (Behaviours.TryGetValue(gameObject, out behaviour))
             {
-                return (T)behaviour;
+                return true;
             }
 
             if (!gameObject.TryGetComponent(out behaviour))
             {
-                behaviour = gameObject.AddComponent<T>();
+                return false;
             }
 
             behaviour.Initialize();
 
-            return (T)behaviour;
+            return true;
         }
 
+        /// <summary>
+        /// Tries to get an existing <see cref="GameObjectBehaviour"/>.
+        /// </summary>
+        /// <returns>True if found a <see cref="GameObjectBehaviour"/>, event if of the wrong type.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static bool TryGetBehaviour<T>(this GameObject gameObject, out T behaviour)
+            where T : GameObjectBehaviour
+        {
+            if (TryGetBehaviour(gameObject, out GameObjectBehaviour value))
+            {
+                behaviour = value as T;
+
+                return true;
+            }
+
+            behaviour = null;
+
+            return false;
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         internal static void AddCachedBehaviour(GameObjectBehaviour behaviour)
         {
             Behaviours.Add(behaviour.CachedGameObject, behaviour);
         }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         internal static void RemoveCachedBehaviour(GameObjectID id)
         {
             Behaviours.Remove(id);